# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API do Sistema de Gestão Escolar - E. M. José Giró Faísca
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login bem-sucedido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Credenciais inválidas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Dados do usuário logado
      responses:
        "200":
          description: Dados do usuário
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Usuario"
        "401":
          description: Não autenticado
...
  /funcionarios/{id}:
    get:
      operationId: getFuncionario
      tags: [funcionarios]
      summary: Perfil de um funcionário
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dados do funcionário
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Funcionario"
...
  /documentos/gerar:
    post:
      operationId: gerarDocumento
      tags: [documentos]
      summary: Gerar documento para um aluno
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GerarDocumentoInput"
      responses:
        "200":
          description: Documento gerado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentoGerado"
  /documentos/listas:
    post:
      operationId: gerarLista
      tags: [documentos]
      summary: Gerar lista de alunos, funcionários ou assinatura
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GerarListaInput"
      responses:
        "200":
          description: Lista gerada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentoGerado"
...
  /arquivo-morto:
    get:
      tags: [arquivo-morto]
      operationId: listarArquivoMorto
      summary: Listar alunos do arquivo morto
      responses:
        "200":
          description: Lista de alunos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ArquivoMortoAluno"
    post:
      tags: [arquivo-morto]
      operationId: criarArquivoMortoAluno
      summary: Adicionar aluno ao arquivo morto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArquivoMortoAlunoInput"
      responses:
        "201":
          description: Aluno criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArquivoMortoAluno"
  /arquivo-morto/{id}:
    get:
      tags: [arquivo-morto]
      operationId: obterArquivoMortoAluno
      summary: Obter aluno do arquivo morto com documentos
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Aluno com documentos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArquivoMortoAlunoDetalhado"
    delete:
      tags: [arquivo-morto]
      operationId: deletarArquivoMortoAluno
      summary: Remover aluno do arquivo morto
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Aluno removido
  /arquivo-morto/{id}/documentos:
    post:
      tags: [arquivo-morto]
      operationId: adicionarDocumentoArquivoMorto
      summary: Adicionar documento ao aluno
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArquivoMortoDocumentoInput"
      responses:
        "201":
          description: Documento adicionado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArquivoMortoDocumento"
  /arquivo-morto/documentos/{docId}:
    delete:
      tags: [arquivo-morto]
      operationId: deletarDocumentoArquivoMorto
      summary: Remover documento
      parameters:
        - name: docId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Documento removido
components:
  schemas:
...
    LoginInput:
      type: object
      properties:
        login:
          type: string
        senha:
          type: string
      required: [login, senha]
    LoginResponse:
      type: object
      properties:
        sucesso:
          type: boolean
        usuario:
          $ref: "#/components/schemas/Usuario"
        mensagem:
          type: string
      required: [sucesso]
    ErrorResponse:
      type: object
      properties:
        erro:
          type: string
        mensagem:
          type: string
      required: [erro, mensagem]
    Usuario:
      type: object
      properties:
        id:
          type: integer
        nomeCompleto:
          type: string
        login:
          type: string
        perfil:
          type: string
          enum: [Master, Direção, Professor]
        genero:
          type: string
      required: [id, nomeCompleto, login, perfil]
...
    Funcionario:
      type: object
      properties:
        id:
          type: integer
        nomeCompleto:
          type: string
        cpf:
          type: string
        funcao:
          type: string
        turno:
          type: string
        telefoneContato:
          type: string
        contatoEmergencia:
          type: string
        dataAdmissao:
          type: string
        vinculo:
          type: string
        status:
          type: string
      required: [id, nomeCompleto]
...
    GerarDocumentoInput:
      type: object
      properties:
        tipoDocumento:
          type: string
        alunoId:
          type: integer
        observacoes:
          type: string
      required: [tipoDocumento, alunoId]
    GerarListaInput:
      type: object
      properties:
        tipo:
          type: string
          enum: [alunos, funcionarios, assinatura_reuniao]
        turmaId:
          type: integer
        titulo:
          type: string
      required: [tipo]
    DocumentoGerado:
      type: object
      properties:
        conteudo:
          type: string
        nomeArquivo:
          type: string
        tipo:
          type: string
      required: [conteudo, nomeArquivo, tipo]
...
    ArquivoMortoAluno:
      type: object
      properties:
        id:
          type: integer
        nomeAluno:
          type: string
        matricula:
          type: string
        anoSaida:
          type: string
        turma:
          type: string
        observacoes:
          type: string
        totalDocumentos:
          type: integer
        criadoEm:
          type: string
      required: [id, nomeAluno]
    ArquivoMortoAlunoInput:
      type: object
      required: [nomeAluno]
      properties:
        nomeAluno:
          type: string
        matricula:
          type: string
        anoSaida:
          type: string
        turma:
          type: string
        observacoes:
          type: string
    ArquivoMortoDocumento:
      type: object
      properties:
        id:
          type: integer
        arquivoMortoId:
          type: integer
        nomeArquivo:
          type: string
        objectPath:
          type: string
        tamanhoBytes:
          type: integer
        criadoEm:
          type: string
      required: [id, arquivoMortoId, nomeArquivo, objectPath]
    ArquivoMortoDocumentoInput:
      type: object
      required: [nomeArquivo, objectPath]
      properties:
        nomeArquivo:
          type: string
        objectPath:
          type: string
        tamanhoBytes:
          type: integer
    ArquivoMortoAlunoDetalhado:
      type: object
      properties:
        id:
          type: integer
        nomeAluno:
          type: string
        matricula:
          type: string
        anoSaida:
          type: string
        turma:
          type: string
        observacoes:
          type: string
        criadoEm:
          type: string
        documentos:
          type: array
          items:
            $ref: "#/components/schemas/ArquivoMortoDocumento"
      required: [id, nomeAluno, documentos]